requirements.txt
Flask==2.0.3
Flask-Cors==3.0.10
Flask-Jsonpify==1.5.0
Flask-RESTful==0.3.9
flask-restplus==0.13.0
Flask-Sockets==0.2.1
pipenv==2022.3.24
python-dateutil==2.8.2
requests
line-bot-sdk==2.4.2
python-dotenv
pymongo==3.8.0
gunicorn
dnspython
gevent==22.8.0
locust==2.12.2
werkzeug==2.0.3
pandas
openpyxl==3.1.5
